fix(nginx): remove nginx build directory
This line was commented out while I was doing some testing of the build
process. Adding this command back in saves about 700MB of disk space. Not
enough that it's a huge concern as I've got plenty on my VPS, but it's enough
to put the commit back in.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>