feat(pagespeed): make pagespeed work
It still isn't the case that pagespeed is a magical panacea for automatically
converting images on the fly or doing other work, but this does start to
take advantage of many of the pagespeed filters to minimize my pages. Of
course, not all of these are needed with HTTP/2, but it's more an exercise
in fun. I'm not quite certain if everything is executing. For example,
I can't really see where images are being rewritten as webp.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>