fix(secure): redirect for Let's Encrypt challenges
This fixes a problem where the Let's Encrypt ACME challenges were hitting the
webdav server which is always protected by a password. Unfortunately, the IETF
ACME spec doesn't allow for these challenges to require a username and
password, so either we need to disable username and password while we do the
auth, disable it for this path, or redirect it somewhere else. I couldn't find
an easy way or Caddy to exclude just one subdirectory from the BasicAuth
handler, so instead I just redirect anything to `.well-known` to my public
server. This solves the problem.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>