feat(vscode): Add VS Code and move to caddy-gen
This is a fairly large change to the repository that migrates away from a
custom set of scripts for templating nginx configuration and toward using
caddy-gen to create custom configruation for caddy on the fly on the basis of
the running docker containers. This results in a much more flexible
configuration that allows me to do a lot more interesting things without
putting the entire stack at risk.

Notably, I've added a Visual Studio Code container to the stack that is served
at a new hostname. This lets me work internally on my machine from a remote
system such as my iPad Pro.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>