chore(cleanup): remove nginx configuration
This removes the old nginx configuration, which was overly complicated and a
little bloated in favor of the new caddy-gen configuration. The whole purpose
of using nginx in the first place was for things like mod_pagespeed and other
tweaks that could really speed up my setup, however, these sort of advancements
are increasingly becoming "normal", so this convoluted and difficult setup is
no longer needed.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>