feat(vscode): add docker support into VS Code
This introduces a couple of changes that make VS code a little nicer.
First, it changes it so it uses the official marketplace - this was needed
because I wanted the Microsoft Remote tools which weren't available in the
coder marketplace.

Secondly, it installs the docker CLI into the VS code container and gives
it access to the host docker instance. Yes, this is definitely insecure,
but it's the best way I could figure out to manage the system. This won't
work in a multi-user world, but it works fine with a single user cloud
machine.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>