feat(hugo): add in live Hugo preview server
One of the things that has kept me from writing as much is that it was usually
too much of a pain to go through and create blog posts. I like Hugo, but it
involved using a Makefile on my local machine and then rsyncing the files up
and realizing that sometimes things wouldn't render properly. This change
creates a server at hugo.wagstrom.net that serves whatever is my working
directory on this machine. In theory, when combined with vscode in the browser,
this should make it so I can write blog articles from anywhere with an LTE
connection.

In the future I'll set it up so this is protected behind an OAuth gateway or
something similar. This will preven the whole world from seeing my current
drafts of future articles.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>